test(test262): consolidate function expression execution tests
