+++ tests/Jroc.Test262.Tests/language/expressions/function/Snapshots/ExecutionTests.dstr_ary_ptrn_elem_ary_rest_init.verified.txt
