+++ tests/Jroc.Test262.Tests/language/expressions/function/Snapshots/ExecutionTests.dstr_ary_ptrn_elem_id_init_fn_name_arrow.verified.txt
