+++ tests/Jroc.Test262.Tests/language/expressions/function/Snapshots/ExecutionTests.dstr_ary_ptrn_elem_id_init_hole.verified.txt
