--- tests/Jroc.Test262.Tests/language/expressions/function/Snapshots/PortAdditionalFunctionCoverageExecutionTests.dflt_params_arg_val_not_undefined.verified.txt
@@ -1,13 +1,0 @@
-﻿true
-true
-true
-true
-true
-true
-true
-true
-true
-true
-true
-true
-true
