--- tests/Jroc.Test262.Tests/language/expressions/function/Snapshots/PortExpressionsBatchExecutionTests.scope_param_rest_elem_var_open.verified.txt
@@ -1,2 +1,0 @@
-﻿true
-true
