--- tests/Jroc.Test262.Tests/language/expressions/function/Snapshots/PortExpressionsBatchExecutionTests.scope_paramsbody_var_close.verified.txt
@@ -1,2 +1,0 @@
-﻿true
-true
