--- tests/Jroc.Test262.Tests/language/expressions/function/Snapshots/PortMoreFunctionDestructuringExecutionTests.dstr_ary_ptrn_elem_ary_elem_iter.verified.txt
@@ -1,4 +1,0 @@
-﻿true
-true
-true
-true
