--- tests/Jroc.Test262.Tests/language/expressions/function/Snapshots/PortMoreFunctionDestructuringExecutionTests.dstr_ary_ptrn_elem_id_iter_done.verified.txt
@@ -1,2 +1,0 @@
-﻿true
-true
